# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Gamics API
  version: 0.9.0
  description: >-
    Gamics player, competition, blind match result verification, Gamics result review, rankings, M-Pesa payments,
    notification, identity and role-based operations API.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/me/sessions/{id}:
    delete:
      operationId: revokeSession
      tags:
        - Sessions
      summary: Revoke one owned session, including the current session.
      description: >-
        Database revocation is propagated through Redis to invalidate outstanding access tokens. Every push
        installation registered with the session is revoked too, so a lost or signed-out phone stops receiving the
        player's pushes.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session and its push installations revoked.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/me/account-deletion:
    get:
      operationId: getAccountDeletionRequest
      tags:
        - Account
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Latest request, or data null when none exists.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/AccountDeletionRequest'
                      - type: 'null'
        '401':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    post:
      operationId: requestAccountDeletion
      tags:
        - Account
      summary: Create or return the pending account-deletion request.
      description: Requests have a 24-hour cooling-off period. Retrying does not restart that period.
      security:
        - bearerAuth: []
      responses:
        '202':
          description: Pending deletion request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDeletionRequest'
        '401':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
    delete:
      operationId: cancelAccountDeletion
      tags:
        - Account
      summary: Cancel the pending request during or after its cooling-off period.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Pending deletion request cancelled.
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/me/account-deletion/execute:
    post:
      operationId: executeAccountDeletion
      tags:
        - Account
      summary: Anonymize the account after the cooling-off period and revoke every session.
      description: >-
        The operation preserves integrity-required competition/payment/audit records under a pseudonymous user row, but
        removes direct profile, contact, device and notification data. It returns 409 if the player owns an organization
        or still participates in an active competition.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - requestId
                - confirmation
              properties:
                requestId:
                  type: string
                  format: uuid
                confirmation:
                  type: string
                  const: DELETE
      responses:
        '204':
          description: Account anonymized and all refresh sessions revoked.
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
...
    AccountDeletionRequest:
      type: object
      additionalProperties: false
      required:
        - id
        - status
        - requestedAt
        - executeAfter
        - cancelledAt
        - completedAt
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - pending
            - cancelled
            - completed
        requestedAt:
          type: string
          format: date-time
        executeAfter:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
...